# apps/backend/src/channels/dto/create-channel.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, IsNotEmpty } from 'class-validator';

import { Channel } from '../schemas/channel.schema';

export class CreateChannelDto
  implements Pick<Channel, 'type' | 'name' | 'token'>
{
  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  type: string;

  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  name: string;

  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  token: string;
}